Add `isPending` to `Computed`
Simply expose the `ReactiveNode` `dirty` property via the new member
function on `Computed`.  Note `ReactiveNode` `dirty` works only for live
consumers which works via a watcher.  This approach will not return true
in some cases that a user might think it will if it is not watched.
This behavior is defined in the tests.